Update readme
Switch away from CDN to local Firebase for easier local debugging, and change $rootScope.$on to $scope.$on since setting up a $rootScope.$on event within a controller is dangerous since that watcher doesn't get automatically destroyed/cleaned up when the controller is destroyed. You could create a $destroy handler, but it's easier to just create the watcher on $scope.